+++ pkg/testcore/suites/helm/helm-suites_test.go
@@ -14,7 +14,7 @@
  *
  */
 
-package suites
+package helm
 
 import (
 	"context"
