+++ pkg/testcore/suites/podprovision/provisioning_test.go
@@ -1,0 +1,240 @@
+package podprovision
+
+import (
+	"context"
+	"fmt"
+	"github.com/dell/cert-csi/pkg/k8sclient"
+	"github.com/dell/cert-csi/pkg/observer"
+	"github.com/dell/cert-csi/pkg/testcore/suites/mockutils"
+	"github.com/stretchr/testify/assert"
+	storagev1 "k8s.io/api/storage/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes/fake"
+	"k8s.io/client-go/rest"
+	"reflect"
+	"testing"
+)
+
+// TestProvisioningSuite_Run
+func TestProvisioningSuite_Run(t *testing.T) {
+	// Create a new context
+	ctx := context.Background()
+
+	// Create a new ProvisioningSuite instance
+	ps := &ProvisioningSuite{
+		VolumeNumber: 1,
+		PodNumber:    1,
+		VolumeSize:   "1Gi",
+		Image:        "quay.io/centos/centos:latest",
+	}
+
+	// Create a fake storage class with VolumeBindingMode set to WaitForFirstConsumer
+	storageClass := &storagev1.StorageClass{
+		ObjectMeta: metav1.ObjectMeta{Name: "test-storage-class"},
+		VolumeBindingMode: func() *storagev1.VolumeBindingMode {
+			mode := storagev1.VolumeBindingWaitForFirstConsumer
+			return &mode
+		}(),
+	}
+
+	// Create a fake k8s clientset with the storage class
+	clientset := fake.NewSimpleClientset(storageClass)
+	clientset = mockutils.MockClientSetPodFunctions(clientset).(*fake.Clientset)
+
+	// Create a fake k8sclient.KubeClient
+	kubeClient := &k8sclient.KubeClient{
+		ClientSet: clientset,
+	}
+
+	// Create PVC client
+	pvcClient, _ := kubeClient.CreatePVCClient("test-namespace")
+
+	// Create Pod client
+	podClient, _ := kubeClient.CreatePodClient("test-namespace")
+
+	// Update the k8sclient.Clients instance with the fake clients
+	clients := &k8sclient.Clients{
+		PVCClient: pvcClient,
+		PodClient: podClient,
+	}
+
+	// Call the Run method
+	_, err := ps.Run(ctx, "test-storage-class", clients)
+	// Check if there was an error
+	if err != nil {
+		t.Errorf("Error running ProvisioningSuite.Run(): %v", err)
+	}
+}
+
+func TestProvisioningSuite_GetObservers(t *testing.T) {
+	ps := &ProvisioningSuite{}
+	obsType := observer.Type("someType")
+	observers := ps.GetObservers(obsType)
+	assert.NotNil(t, observers)
+	// Add more assertions based on expected behavior
+}
+
+func TestProvisioningSuite_GetClients(t *testing.T) {
+	client := fake.NewSimpleClientset()
+
+	kubeClient := k8sclient.KubeClient{
+		ClientSet:   client,
+		Config:      &rest.Config{},
+		VersionInfo: nil,
+	}
+
+	pvcClient, _ := kubeClient.CreatePVCClient("test-namespace")
+	podClient, _ := kubeClient.CreatePodClient("test-namespace")
+	vaClient, _ := kubeClient.CreateVaClient("test-namespace")
+	metricsClient, _ := kubeClient.CreateMetricsClient("test-namespace")
+
+	type args struct {
+		namespace string
+		client    *k8sclient.KubeClient
+	}
+	tests := []struct {
+		name    string
+		ps      *ProvisioningSuite
+		args    args
+		want    *k8sclient.Clients
+		wantErr bool
+	}{
+		{
+			name: "Testing GetClients",
+			ps:   &ProvisioningSuite{},
+			args: args{
+				namespace: "test-namespace",
+				client:    &kubeClient,
+			},
+			want: &k8sclient.Clients{
+				PVCClient:     pvcClient,
+				PodClient:     podClient,
+				VaClient:      vaClient,
+				MetricsClient: metricsClient,
+			},
+			wantErr: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := tt.ps.GetClients(tt.args.namespace, tt.args.client)
+			fmt.Println(got, err)
+
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ProvisioningSuite.GetClients() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			fmt.Println(reflect.TypeOf(got), reflect.TypeOf(tt.want))
+			if reflect.TypeOf(got) != reflect.TypeOf(tt.want) {
+				t.Errorf("ProvisioningSuite.GetClients() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestProvisioningSuite_GetNamespace(t *testing.T) {
+	tests := []struct {
+		name string
+		ps   *ProvisioningSuite
+		want string
+	}{
+		{
+			name: "Testing GetNamespace",
+			ps:   &ProvisioningSuite{},
+			want: "prov-test",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := tt.ps.GetNamespace(); got != tt.want {
+				t.Errorf("ProvisioningSuite.GetNamespace() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestProvisioningSuite_GetName(t *testing.T) {
+	tests := []struct {
+		name string
+		ps   *ProvisioningSuite
+		want string
+	}{
+		{
+			name: "Testing GetName when Description is not empty",
+			ps: &ProvisioningSuite{
+				Description: "test-description",
+			},
+			want: "test-description",
+		},
+		{
+			name: "Testing GetName when Description is empty",
+			ps:   &ProvisioningSuite{},
+			want: "ProvisioningSuite",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := tt.ps.GetName(); got != tt.want {
+				t.Errorf("ProvisioningSuite.GetName() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestProvisioningSuite_Parameters(t *testing.T) {
+	tests := []struct {
+		name string
+		ps   *ProvisioningSuite
+		want string
+	}{
+		{
+			name: "Testing Parameters",
+			ps: &ProvisioningSuite{
+				PodNumber:    1,
+				VolumeNumber: 5,
+				VolumeSize:   "3Gi",
+			},
+			want: "{pods: 1, volumes: 5, volumeSize: 3Gi}",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := tt.ps.Parameters(); got != tt.want {
+				t.Errorf("ProvisioningSuite.Parameters() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestProvisioningSuite_validateCustomPodName(t *testing.T) {
+	tests := []struct {
+		name string
+		ps   *ProvisioningSuite
+		want string
+	}{
+		{
+			name: "Testing validateCustomPodName with single pod and custom name",
+			ps: &ProvisioningSuite{
+				PodNumber:     1,
+				PodCustomName: "custom-pod",
+			},
+			want: "custom-pod",
+		},
+		{
+			name: "Testing validateCustomPodName with multiple pods",
+			ps: &ProvisioningSuite{
+				PodNumber:     2,
+				PodCustomName: "custom-pod",
+			},
+			want: "",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tt.ps.validateCustomPodName()
+			if got := tt.ps.PodCustomName; got != tt.want {
+				t.Errorf("ProvisioningSuite.validateCustomPodName() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
